Cleanup RawImageCreator a bit
This patch cleans-up the RawImageCreator code a little bit. We construct
the full name for the files we generate many times in various places, which
looks untidy. This commit introduces 2 helper functions: _full_name() and
_full_path() which return full name for a file and full path to a file,
correspondingly. The code becomes more readable with this change.

But also, this is a preparation to one of the following patches where we'll
need to construct full path for the bmap file. We do not want to add yet
another place where we construct this manually, so we better have a helper
function.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>
Change-Id: I9ac52c95e4b16bcd227df467439ac2ae87b1c145