RawImageCreator: use hashlib unconditionally
mic does not work with python < 2.5 (see distfiles/mic.spec, for example), so
we can use hashlib without falling-back to 'import sha1'. This patch cleans-up
the code a bit and removes the fall-back to 'import sha1' stuff.

I move 'import hashlib' to the top level because it is cleaner and because I
will need hashlib very soon in the next patch.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>