workaround to fix loop device releasing
now mic tries to save process pid to a file named
with the same name if the device used, and clean
up all devices according to those like files before
finding a new loop device.

it's easy to detect if the loop device used or not
by checking its content, the process pid. if the
process with this pid is alive, don't clean up, else
do it

Signed-off-by: Gui Chen <gui.chen@intel.com>