ExtDiskMount: fix the UUID feature for extX
The current implementation of ExtDiskMount is strange. Instead of generating a
random UUID and then ask mkfs.extX to use that UUID, it run mkfs.extX without
-U, let's mkfs.extX generate a random UUID, and then uses e2fsdump and parses
its output.

This is not very logical, and this also does not work with the version of
mkfs that we use at the moment.

Change the logic and simply use mkfs.extX -U.

Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>