kickstart: add the installerfw KS command
This command will be used in the KS file and it will specify the list legacy
MIC features which have to be disabled. Instead of these feature, the general
installer framework infrastructure mechanisms will be used.

The first option will be "extlinux" which will disable the legacy extlinux
installation feature. At some point, installerfw=all will be supported,
it will disable all the legacy features (for keyboard, fstab, users, etc).

Along with the option, introduce a helper function which checks if a the
installer framework has to be used for a feature. Usage example:

if use_installerfw(ks, "extlinux"):
    # do not execute MIC's built-in extlinux installation code
else:
    # the compatibility legacy mode

original written by Artem
Artem: added the use_installerfw() function.

Signed-off-by: Gui Chen <gui.chen@intel.com>
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>