baseimager: export partition device node
Also export the device node name for each partition. This will be helpful when
we start generating fstab entries in the installer framework scripts.

In MIC environment, the device node name is easily constructed from the disk
name.  The disk name compes from the KS file from the --ondisk option of the
'part' command. And --ondisk is a mandatory argument.

In other installer environments, the disk names may be very different.

Change-Id: Idafce950e511f4a7693e17be10e69fbb1e7d00d8
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>