liveusb: fix vfat UUID
Commit '3577b518715c29175c5f517c6f3cf7a70b0e3d39' fixed the UUID string for
raw images, but broke it for liveusb images. And this patch fixes the issue by
making the liveusb code stop adding a dash to the UUID, because now the dash is
there, just like in UUID of ext4.

Change-Id: Ibfccdf204b301808ef63589dbc3934e517a0212d
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>