kickstart: add an alias for installerfw
In commit "d52080c kickstart: add the installerfw KS command" we introduced the
"installerfw" option. However, the later feed-back was that the name is very
confusing and non-intuitive, and we agreed to re-name this option to
"installerfw_plugins". With this name it is more obvious that this command
enables a list of plugins.

The name of the "extlinux" option of this command was also criticized, and we
agreed to rename it to more generic "bootloader" name.

We agreed that we do the following.

1. Add "installerfw_plugins" alias for "installerfw"
2. Add "bootloader" alias for "extlinux"
3. When users use the old "installerfw" and "extlinux" names, accept this,
   but print a warning which says that these obsolete names will be removed in
   future releases
4. Remove the old keywords later, after a couple releases.

This patch implements just that - introduces the aliases, prints warnings, makes
sure that internally we only use the new names, amends commentaries.

Base on suggestions and the initial patch from Alexander Kanevskiy.

Change-Id: I4646d437833f5252061cb963385269dc240f5792
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>