raw.py: create device nodes to access own disk
This patch creates loop-back devices in /dev of the image so that it would be
possible to access own disk from within chroot environment. This makes it
possible to install extlinux from chroot, not outside of chroot, as it is done
now. This also allows to do extlinux installation in an installer framework
plugin.

I do realize that this is not the most beautiful thing to do, but it works - I
have a script which installs extlinux when I run it from %post.

Change-Id: I2b14962a37524a15c5dc493572ae8e91ca8523ae
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>