kickstart: remove old installerfw keywords
The story: the installer framework keyword in KS files originally was
"installerfw", and the keyword for the first plugin was "extlinux". However,
we realized that this is bad naming, and commit changed the names in commit
'0bafaa1b31a0 kickstart: add an alias for installerfw'. That commit preserved
the old name as aliases for compatibility reasons. And added scary warnings
which are printed when the old names are used.

The only user user of the old names was Tizen IVI, and it stopped using them
very long time ago. Now it uses the new 'installerfw_plugins' and 'bootloader'
keywords instead.

So it is time to kill the old keywords completely and clean the code up. And
this is exactly what this patch does.

Change-Id: I23ef6c5850744734816974233229cf99fca05f38
Signed-off-by: Artem Bityutskiy <artem.bityutskiy@intel.com>